modules/kafka-producer/app/requirements.txt
confluent-kafka==2.13.2
python-dotenv